+++ Source/FieldSolver/SpectralSolver/Make.package
@@ -1,0 +1,8 @@
+CEXE_headers += WarpXComplex.H
+CEXE_headers += SpectralData.H
+CEXE_sources += SpectralData.cpp
+CEXE_headers += PsatdSolver.H
+CEXE_sources += PsatdSolver.cpp
+
+INCLUDE_LOCATIONS += $(WARPX_HOME)/Source/FieldSolver/SpectralSolver
+VPATH_LOCATIONS   += $(WARPX_HOME)/Source/FieldSolver/SpectralSolver
